+++ DTModelStorageTests/XCTests/Helpers/ListItem+CoreDataProperties.swift
@@ -1,0 +1,19 @@
+//
+//  ListItem+CoreDataProperties.swift
+//  DTModelStorage
+//
+//  Created by Denys Telezhkin on 29.10.15.
+//  Copyright © 2015 Denys Telezhkin. All rights reserved.
+//
+//  Choose "Create NSManagedObject Subclass…" from the Core Data editor menu
+//  to delete and recreate this implementation file for your updated model.
+//
+
+import Foundation
+import CoreData
+
+extension ListItem {
+
+    @NSManaged var value: NSNumber
+
+}